requirements.txt
fastapi[standard]==0.111.0
uvicorn[standard]==0.29.0
pypdf==5.2.0
langchain==0.2.15
langchain_community==0.2.14
langchain-google-genai==1.0.10
python-dotenv
python-socketio
PyJWT